refactor(moirai): Retire NUMA helper
