+++ Cargo.toml
@@ -23,7 +23,7 @@ members = [
 resolver = "2"
 
 [workspace.package]
-version = "0.2.0"
+version = "0.3.0"
 edition = "2021"
 authors = ["Moirai Team"]
 license = "MIT OR Apache-2.0"
@@ -36,21 +36,21 @@ rust-version = "1.75.0"
 
 [workspace.dependencies]
 # Internal dependencies
-moirai-core = { path = "moirai-core", version = "0.2.0" }
-moirai-executor = { path = "moirai-executor", version = "0.2.0" }
-moirai-scheduler = { path = "moirai-scheduler", version = "0.2.0" }
-moirai-transport = { path = "moirai-transport", version = "0.2.0" }
-moirai-sync = { path = "moirai-sync", version = "0.2.0" }
-moirai-async = { path = "moirai-async", version = "0.2.0" }
-moirai-async-macros = { path = "moirai-async-macros", version = "0.2.0" }
-moirai-pal = { path = "moirai-pal", version = "0.2.0" }
-moirai-iter = { path = "moirai-iter", version = "0.2.0" }
-moirai-parallel = { path = "moirai-parallel", version = "0.2.0" }
-moirai-tls = { path = "moirai-tls", version = "0.2.0" }
-moirai-http = { path = "moirai-http", version = "0.2.0" }
-moirai-metrics = { path = "moirai-metrics", version = "0.2.0" }
-moirai-utils = { path = "moirai-utils", version = "0.2.0", features = ["std"] }
-moirai-gpu = { path = "moirai-gpu", version = "0.2.0" }
+moirai-core = { path = "moirai-core", version = "0.3.0" }
+moirai-executor = { path = "moirai-executor", version = "0.3.0" }
+moirai-scheduler = { path = "moirai-scheduler", version = "0.3.0" }
+moirai-transport = { path = "moirai-transport", version = "0.3.0" }
+moirai-sync = { path = "moirai-sync", version = "0.3.0" }
+moirai-async = { path = "moirai-async", version = "0.3.0" }
+moirai-async-macros = { path = "moirai-async-macros", version = "0.3.0" }
+moirai-pal = { path = "moirai-pal", version = "0.3.0" }
+moirai-iter = { path = "moirai-iter", version = "0.3.0" }
+moirai-parallel = { path = "moirai-parallel", version = "0.3.0" }
+moirai-tls = { path = "moirai-tls", version = "0.3.0" }
+moirai-http = { path = "moirai-http", version = "0.3.0" }
+moirai-metrics = { path = "moirai-metrics", version = "0.3.0" }
+moirai-utils = { path = "moirai-utils", version = "0.3.0", features = ["std"] }
+moirai-gpu = { path = "moirai-gpu", version = "0.3.0" }
 melinoe = { git = "https://github.com/ryancinsight/melinoe.git", rev = "bb07447f6c71978c6fed47114ba9ecfda965ddc9", version = "0.9.0", default-features = false }
 themis = { git = "https://github.com/ryancinsight/themis", rev = "6140468c79279ec8f112641ea7422cef4688c7f6", version = "0.10", default-features = false, features = ["std"] }
 # Kernel resource budget vocabulary for the GPU occupancy planner (atlas ADR 0002).
