+++ benchmarks/Cargo.toml
@@ -114,10 +114,6 @@ name = "execution_context_comparison"
 harness = false
 
 [[bench]]
-name = "numa_context_comparison"
-harness = false
-
-[[bench]]
 name = "async_iterator_comparison"
 harness = false
