+++ benchmarks/tests/benchmark_contracts/artifact_contracts.rs
@@ -201,10 +201,10 @@ fn version_artifacts_are_synchronized_for_current_target() {
     let changelog = read_benchmark("../CHANGELOG.md");
 
     for required in [
-        "version = \"0.2.0\"",
-        "moirai-core = { path = \"moirai-core\", version = \"0.2.0\" }",
-        "moirai-executor = { path = \"moirai-executor\", version = \"0.2.0\" }",
-        "moirai-iter = { path = \"moirai-iter\", version = \"0.2.0\" }",
+        "version = \"0.3.0\"",
+        "moirai-core = { path = \"moirai-core\", version = \"0.3.0\" }",
+        "moirai-executor = { path = \"moirai-executor\", version = \"0.3.0\" }",
+        "moirai-iter = { path = \"moirai-iter\", version = \"0.3.0\" }",
     ] {
         assert!(
             manifest.contains(required),
@@ -213,12 +213,12 @@ fn version_artifacts_are_synchronized_for_current_target() {
     }
 
     assert!(
-        checklist.contains("**Target Version**: 0.2.0"),
+        checklist.contains("**Target Version**: 0.3.0"),
         "checklist target version must match Cargo workspace version"
     );
     assert!(
-        changelog.contains("## [0.2.0] - 2026-05-24"),
-        "changelog must contain a synchronized 0.2.0 section"
+        changelog.contains("## [0.3.0] - 2026-07-15"),
+        "changelog must contain a synchronized 0.3.0 section"
     );
 }
