+++ docs/checklist.md
@@ -2,12 +2,21 @@
 
 ## Current State: Unified Scheduler Sprint
 
-**Target Version**: 0.2.0
+**Target Version**: 0.3.0
 **Codebase Status**: Executor scheduler refactor complete; `HybridRouter<P>` provides sealed ZST thread/process/server/accelerator/async-lane route policies, `moirai-transport` consumes selected process/server routes through fixed-format remote task execution, the top-level `Moirai` facade exposes only fixed-capability routed process/server execution, and Mnemosyne-owned archive bytes define the current route ownership handoff across thread/process/server/device regions. Scoped scheduler batches exceed Tokio/Rayon ready-work benchmark baselines; indexed map/reduce matches normalized Rayon fixed-pool indexed CPU work and uses worker-plus-caller chunk caps; mixed sync completion, async result, and indexed reduction work has a value-checked unified-scheduler comparison against a Tokio plus Rayon two-engine reference; real-application mixed workload coverage combines async fan-out, scoped request work, indexed analytics, bounded channel transfer, and closed-form checksums against a Tokio plus Rayon reference; public per-task handles use atomic result slots with a monomorphized zero-sized wait policy, direct first CAS, load-gated pending spins, and state-machine-gated inline waiter cells; async public handles use inline future/lifecycle/result-sender storage with wake-coalesced inline repoll and direct future-state wakers; iterator and scheduler hot paths use monomorphized erased storage at heterogeneous boundaries instead of `dyn` execution dispatch; transport safe-channel payloads use rkyv-style archive views over owned bytes; explicit Rayon/Tokio gap audit tracks executable comparison coverage, same-run benchmark evidence, runtime dependency boundaries, and bounded ecosystem parity.
 **Architecture**: Unified scheduler/router for local CPU worker threads, sync/blocking/async-ready work classes, process routes, server routes, accelerator metadata routes, and per-process async lanes; device-memory regions and accelerator backend consumption remain open architecture items.
 **Quality Level**: Focused scheduler/core/PAL/benchmark clippy clean; scoped, indexed, industry, and public-handle Criterion targets pass with value assertions under bounded verification
 
-**Current verification**: Rust 1.97 workspace format and warning-denied all-target/all-feature Clippy pass; canonical all-feature nextest passes 722/722 in 8.1 seconds; doctests and warning-clean workspace rustdoc pass. The bounded scheduler admission benchmark contract verifies fallible queue publication and explicit capacity handling. Themis is pinned to audited revision `6140468c79279ec8f112641ea7422cef4688c7f6`, and Mnemosyne is pinned to provider-identity commit `7435ea0b305210874d6f3faae7860f40de835426`.
+**Current verification**: focused 0.3.0 formatter, warning-denied all-target/all-feature Clippy, rustdoc, and nextest gates pass for `moirai-core`, `moirai-iter`, and `moirai-benchmarks` (323 tests, 2 cfg-skips). The bounded scheduler admission benchmark contract verifies fallible queue publication and explicit capacity handling. Themis is pinned to audited revision `6140468c79279ec8f112641ea7422cef4688c7f6`, and Mnemosyne is pinned to provider-identity commit `7435ea0b305210874d6f3faae7860f40de835426`.
+
+### MOI-NUMA-001 / MOI-TREE-001 closure
+
+- [x] [major] Delete the unconsumed `moirai_iter::numa` API, its benchmark,
+  and its source contract. Themis owns placement, Mnemosyne owns allocation,
+  and `moirai-parallel` owns scheduler-backed data-parallel execution.
+- [x] [patch] Split hybrid and MPMC channel logic into endpoint/state leaves.
+  `HybridChannel<T>` is now a zero-sized factory; endpoint halves own the live
+  synchronization state.
 
 ---
